rasseiver: read the configuration file in main.cpp
This adds the ability for the user to pass a config file as an argument, and to
use it to configure Rasseiver.

If the user specifies "airspy" as the device, it will try to open an Airspy with
the specified configuration.  Otherwise, it will fail as we do not support any
other device yet.

Signed-off-by: Alban Gruin <alban@pa1ch.fr>