rasseiver: add a circular buffer implementation
This adds a circular buffer implementation.  This will be useful for filtering
operations, as we have to keep older values in order to properly filter these
values.

The implementation is based on a templated class using two std::vector<>s to
store its data; one for the "previous" values, and another for the "current"
values.

A swap operation is provided to exchange their value and clear the current
buffer.  An naive technique to swap them would be to call std::move(), but the
memory region contained by the previous buffer may be freed during this
operation.  To avoid doing free/alloc cycles, they are swapped with std::swap()
instead, and the current buffer is manually cleared.

Signed-off-by: Alban Gruin <alban@pa1ch.fr>