rasseiver/CircularBuffer: allocate a specified amount of memory
This adds a constructor that allocates a specified amount of memory.

This avoids allocating memory during the processing of the buffer, which could
be as reliable and fast as possible.  Dynamic memory allocation takes time and
can fail on some systems.

Doing this at initialisation allows the software to crash at startup, and even
if annoying, it's infinitely less than in the middle of a work session.

Signed-off-by: Alban Gruin <alban@pa1ch.fr>