rasseiver: stop when sigint or sigterm is received
Until now, there was no clean way to stop the program.  This adds a way to do so
properly, by sending a sigint or a sigterm to the program.

As the program is multi-threaded, we have to mask the two signals we want to
handle with pthread_sigmask(3), and then wait for them with sigwait(3).

Signed-off-by: Alban Gruin <alban@pa1ch.fr>