rasseiver/filter: add an implementation of an FIR filter
This adds an implementation of an FIR filter.  The function computes the
convolution between the input and the filter, and inserts the result in a
std::vector<>.

This implementation also can perform a decimation on the input buffer by filtering
only the values that will be copied to the output, and should be faster
than filtering the entire input, and then decimating the result.

Signed-off-by: Alban Gruin <alban@pa1ch.fr>