rasseiver: add a `Process' functor
As part of the program, we need to filter and decimate the data received from
the Airspy (or another device).  With the Airspy library, we have to call a
function to start streaming, and each time a buffer is ready, this function will
call a callback.  As the Airspy library is written in C, it cannot call lambdas
or functors itself.

This adds a class that stores a circular buffer, a filter, and an output, that
can apply different processes to an input provided.

This is achieved through a custom functor, instead of a lambda function or a
more traditionnal function pointer.

For portability reasons (the library for Airspy does not necessarily uses the
same structures as another), an intermediary function call is needed.

If we use correctly the functor (ie. using templates), it may end up performing
better than a "simple" call.  Benchmarking is required here.

Signed-off-by: Alban Gruin <alban@pa1ch.fr>