rasseiver/device: templated device, use Process in devices
This makes Device templated (eg. we want float or int16_t values), and receive
requests a Process.

Airspy only supports int16_t for now (needs work).  It calls the function to
start streaming from the airspy, add a callback to call the Process functor, and
closes the stream.

This also creates the associated process in main.cpp for Airspy devices.

Signed-off-by: Alban Gruin <alban@pa1ch.fr>