rasseiver/process: replace operator() by a more traditionnal method
The operator() method is not very readable for C++ beginners and not that useful
-- it's not used as a functor yet, and worse, the operator is called by its name
and not through an operator:

    auto *process = static_cast<Process<T> *> (transfer->ctx);
    process->operator()(…);

This replaces this operator by a traditionnal method, which is more readable
than operator() in our case.

This also removes the usage of std::iota in the constructor of Process.  The
array is constructed with a list.

Signed-off-by: Alban Gruin <alban@pa1ch.fr>