rasseiver/filter: store temporary values in double, not T
Until now, values being filtered are stored in a variable of type T, whereas the
filter uses doubles.

This has two drawbacks.  First, converting the values from double to T,
especially when T is an integer type, induces a huge loss of accuracy with each
iteration of the filter.  Second, the results needs to be converted at each
iteration, instead of being converted once at the end.

This changes the type of the temporary values to doubles, and converts them only
when they are inserted into the output buffer.  This improves the performance of
the filter by a factor of 4 on an amd64 computer, and by a factor of 2.5 on a
Raspberry Pi 1B+.

Another part that may be optimised: for each iteration, many values from the
input buffer have to be converted from an integer to a double, and many of them
will be converted as much as the filter size.  Perhaps they could be stored in a
buffer of double to avoid converting them over and over again.

Signed-off-by: Alban Gruin <alban@pa1ch.fr>