rasseiver/device: receiving data is now a ressource
Sometimes, when asking rasseiver to stop, it segfaults instead of exiting
gracefully.  This is because the Process was still filtering data from the
Airspy, reading and writing in the buffers of Process.  The receipt of data can
be stopped by deleting the Airspy, which calls the process in a thread.

Unfortunately, Process needs an information (the size of the buffer) from the
Airspy, so it must be constructed after the Airspy.  This means it will be
destructed before the Airspy, with its buffers.  If the process was still
filtering the buffers, an invalid read or write will follow, ultimately leading
to a segmentation fault.

To fix this, the receipt of data must stop before the destruction of the
Process.  Receiving data can be seen as a ressource to be acquired after the
creation of the Process, and released before its destruction.

A new class, Receiver, is added to fulfill this role.  Its constructor starts
the receipt of data, its destructor stops it.  This class is generic and can
handle any kind of device, and any flavour of Process<T>.  Devices now needs to
define a method, stop(), to stop the data reception.

run_airspy() is modified to use of Receiver<Airspy, int16_t>, and some comments
are added.  The Receiver is constructed after the Process, and should be
destroyed before, so that the filtering process should have stopped before the
buffers are deallocated.  This should fix the segfaults.

Signed-off-by: Alban Gruin <alban@pa1ch.fr>