rasseiver/filter: detect saturation
This adds a saturation detector to the filter.  The detection is very simple:
when the modulus of a sample exceeds a threshold value, it means there is
saturation.  This method is very crude, but fast nonetheless.

Instead of returning the index of the loop, filter_buffer() returns whether it
is saturated or not.  The index is modified with a reference instead.

The threshold used in the filter is stored in a Process, which obtains it from
its constructor.  The value passed there is multiplied by 92% before being
stored.

The process created in run_airspy() uses the maximum value that an airspy can
sample as a threshold.

Signed-off-by: Alban Gruin <alban@pa1ch.fr>