server/sender: better file descriptor and socket management
This adds a RAII wrapper to manage file descriptors properly.  When initialised,
it acquires a file descriptor, and it closes it when it is destructed.  It may
be marked as unusable, in which case the file descriptor should not be used.

This RAII wrapper is then used in the Sender class.  It allows a better
management of ressources during the connection.  The Sender class stores the
address and the port of the server to facilitate reconnection.

reconnect() does not throw errors anymore.  Instead, it returns 0 if the
connection succeeded, or -1 if not.  In this case, the constructor throws an
error.  This is done to avoid handling exceptions during the process, which can
be more expensive than return values.

Signed-off-by: Alban Gruin <alban@pa1ch.fr>