rasseiver/sender: send a header before the actual data
This adds a header to the packets sent on the network.

The header contains the amount of bytes to be sent on the network in its first 8
bytes, and some flags in another byte.  Currently, there is only one flag,
located on the LSB, indicating whether the data is saturated or not.

Signed-off-by: Alban Gruin <alban@pa1ch.fr>