ilserver: add the notion of channels
This add the notion of channels to each server.  For each server, there can only
be one channel, and so one connection.

The channel stores its level, offset, buffer, synchronisation state, etc.

To synchronise a channel with the reference channel, a peak in current must be
injected in the reference channels and in the channels to synchronise.  This
way, it can be synchronised in time, but not in intensity.

Signed-off-by: Alban Gruin <alban@pa1ch.fr>