ilserver: compute the median to synchronise the levels
To synchronise the levels, a simple ratio between the last modulus of the
current channel and that of the reference channel was used, which is not very
precise.

This changes this to compute the medians of the moduli after the start of the
carrier.

The value written in the CSV is multiplicated by the resulting ratio.

Signed-off-by: Alban Gruin <alban@pa1ch.fr>