ilserver: complete rewrite of channels
This completes the rewriting of the channel system.  The second buffer and the
position indication are deleted.

The buffers are now written to a single file.  This happens when the reference
channel receives a packet.  The amount of samples to be written is the minimum
number of samples available all channels.  If a channel is not initialised or
not synchronised, no samples is written.

Signed-off-by: Alban Gruin <alban@pa1ch.fr>