ilserver/channel: remove the mutex from the reference channel
This removes the mutex of the reference channel.

The mutex was from asyncio, which means it's not guaranteed to be
thread-safe, which will be important when adding threading support.
Moreover, it's unnecessary, since all operations on channels are made
by the same thread -- currently, the asyncio thread, soon to be a
separate, but still unique thread, and functions acquiring the mutex
must be asynchronous, meaning they can't be called from synchronous
code (ie. from a non-asyncio thread.)

Signed-off-by: Alban Gruin <alban@pa1ch.fr>