ilserver: delegate buffer processing to a thread
This adds a thread pool of one thread to handle the processing of the
channels.  As there is only one thread, there is no risk of race
condition between the channels.

In addition, the csv writer is dropped, and the processing is splitted
in group of PACKET_SIZE samples, and compensate is modified to work on
lists.

Signed-off-by: Alban Gruin <alban@pa1ch.fr>