ilserver: add a controller server
This adds a controller server to control the phase shift to be applied
to each channel.  Each processed packet is then sent to the controller.

When the controller connects, the server sends the number of channels.

Each packet is prepended by a boolean to indicate whether or not the
phase shift was updated for that packet.

For now, it is not safe to have multiple controllers, but this should be
addressed in the future.

Signed-off-by: Alban Gruin <alban@pa1ch.fr>