rasseiver/filter: optimise the filter
This optimises the filter by splitting the filtering operation of a
single sample in two parts: one using samples in the previous buffer, if
necessary, and the other using samples in the current buffer.  This also
simplifies the calculation of indices.

This reduces the filtering time on a Raspberry Pi 3B (~22ms vs ~9ms
using gcc 6, -O2 -march=native -flto).

Other ways were investigated (eg. starting from the end of the filter
instead of the beginning), but they were slower than the one used in
this commit.

Signed-off-by: Alban Gruin <alban@pa1ch.fr>