libils: add functions to receive a large amount of data from a socket
According to the Python documentation[0], the number of bytes to read
from a socket should be a small power of two, like 4096.  This
introduces async_recv_data() and recv_data(), two helpers functions to
efficiently receive large amounts of data from a socket, asynchronously
or synchronously, by steps of up to 4096.

Functions to receive data in the base receiver class are two now: one to
just receive the data (receive()), and another to wait for the shifts to
be applied (wait_and_receive()).

[0] https://docs.python.org/3/library/socket.html#socket.socket.recv

Signed-off-by: Alban Gruin <alban@pa1ch.fr>