rasseiver/device_airspy: add a macro to call a libairspy setter
Some libairspy functions acting on a device may fail (eg. opening a
device, setting the sample rate, etc.)  Our code has wrappers calling
them, throwing an exception in case of error.  They all looked alike.

This extracts all these functions to a macro to reduce code duplication.
They are not extracted as a function to be able to put the exact
function name in the exception, without typing it twice, which is
impossible without stringification.

Signed-off-by: Alban Gruin <alban@pa1ch.fr>